move WriteRaceChecker to loopy.schedule.tools
