+++ loopy/schedule/tools.py
@@ -21,6 +21,8 @@ THE SOFTWARE.
 """
 
 from loopy.kernel.data import AddressSpace
+from pytools import memoize_method
+import islpy as isl
 
 
 # {{{ block boundary finder
@@ -136,5 +138,217 @@ def get_return_from_kernel_mapping(kernel):
             raise NotImplementedError(type(sched_item))
 
     return return_from_kernel_idxs
+
+# }}}
+
+
+# {{{ check for write races in accesses
+
+def _check_for_access_races(map_a, insn_a, map_b, insn_b, knl, callables_table):
+    """
+    Returns *True* if the execution instances of *insn_a* and *insn_b*, accessing
+    the same variable via access maps *map_a* and *map_b*, result in an access race.
+
+    .. note::
+
+        The accesses map_a, map_b lead to write races iff there exists 2
+        *unequal* global ids that access the same address.
+    """
+    import pymbolic.primitives as p
+    from loopy.symbolic import isl_set_from_expr
+    from loopy.kernel.data import (filter_iname_tags_by_type,
+                                   HardwareConcurrentTag)
+
+    gsize, lsize = knl.get_grid_size_upper_bounds(callables_table,
+                                                  return_dict=True)
+
+    # {{{ Step 1: Preprocess the maps
+
+    # Step 1.1: Project out inames which are also map's dims, but does not form the
+    #           insn's within_inames
+    # Step 1.2: Project out sequential inames in the access maps
+    # Step 1.3: Rename the dims with their iname tags i.e. (g.i or l.i)
+    # Step 1.4: Name the ith output dims as _lp_dim{i}
+
+    updated_maps = []
+
+    for (map_, insn) in [
+            (map_a, insn_a),
+            (map_b, insn_b)]:
+        for iname in ((set(map_.get_var_dict().keys()) & knl.all_inames())
+                      - insn.within_inames):
+            # project out all inames in the map's domain that aren't part of the
+            # insn's domain
+            dt, pos = map_.get_var_dict()[iname]
+            map_ = map_.project_out(dt, pos, 1)
+
+        for iname in insn.within_inames:
+            if not filter_iname_tags_by_type(knl.inames[iname].tags,
+                                             HardwareConcurrentTag):
+                dt, pos = map_.get_var_dict()[iname]
+                map_ = map_.project_out(dt, pos, 1)
+
+        for name, (dt, pos) in map_.get_var_dict().items():
+            if dt == isl.dim_type.in_:
+                tag, = filter_iname_tags_by_type(knl.inames[name].tags,
+                                                 HardwareConcurrentTag)
+                map_ = map_.set_dim_name(dt, pos, str(tag))
+
+        for i_l in lsize:
+            if f"l.{i_l}" not in map_.get_var_dict():
+                ndim = map_.dim(isl.dim_type.in_)
+                map_ = map_.add_dims(isl.dim_type.in_, 1)
+                map_ = map_.set_dim_name(isl.dim_type.in_, ndim, f"l.{i_l}")
+
+        for i_g in gsize:
+            if f"g.{i_g}" not in map_.get_var_dict():
+                ndim = map_.dim(isl.dim_type.in_)
+                map_ = map_.add_dims(isl.dim_type.in_, 1)
+                map_ = map_.set_dim_name(isl.dim_type.in_, ndim, f"g.{i_g}")
+
+        for pos in range(map_.dim(isl.dim_type.out)):
+            map_ = map_.set_dim_name(isl.dim_type.out, pos, f"_lp_dim{pos}")
+
+        updated_maps.append(map_)
+
+    map_a, map_b = updated_maps
+
+    # }}}
+
+    # {{{ Step 2: rename all lid's, gid's in map_a to lid.A, gid.A
+
+    for name, (dt, pos) in map_a.get_var_dict().items():
+        if dt == isl.dim_type.in_:
+            map_a = map_a.set_dim_name(dt, pos, name+".A")
+
+    # }}}
+
+    # {{{ Step 3: rename all lid's, gid's in map_b to lid.B, gid.B
+
+    for name, (dt, pos) in map_b.get_var_dict().items():
+        if dt == isl.dim_type.in_:
+            map_b = map_b.set_dim_name(dt, pos, name+".B")
+
+    # }}}
+
+    # {{{ Step 4: make map_a, map_b ISL sets
+
+    map_a, map_b = isl.align_two(map_a, map_b)
+    map_a = map_a.move_dims(isl.dim_type.in_, map_a.dim(isl.dim_type.in_),
+                            isl.dim_type.out, 0, map_a.dim(isl.dim_type.out))
+
+    map_b = map_b.move_dims(isl.dim_type.in_, map_b.dim(isl.dim_type.in_),
+                            isl.dim_type.out, 0, map_b.dim(isl.dim_type.out))
+    set_a = map_a.domain()
+    set_b = map_b.domain()
+
+    # }}}
+
+    assert set_a.get_space() == set_b.get_space()
+
+    # {{{ Step 5: create the set any(l.i.A != l.i.B) OR any(g.i.A != g.i.B)
+
+    space = set_a.space
+    unequal_global_id_set = isl.Set.empty(set_a.get_space())
+
+    for i_l in lsize:
+        lid_a = p.Variable(f"l.{i_l}.A")
+        lid_b = p.Variable(f"l.{i_l}.B")
+        unequal_global_id_set |= (isl_set_from_expr(space,
+                                                    p.Comparison(lid_a, "!=", lid_b))
+                                  )
+
+    for i_g in gsize:
+        gid_a = p.Variable(f"g.{i_g}.A")
+        gid_b = p.Variable(f"g.{i_g}.B")
+        unequal_global_id_set |= (isl_set_from_expr(space,
+                                                    p.Comparison(gid_a, "!=", gid_b))
+                                  )
+
+    # }}}
+
+    return not (set_a & set_b & unequal_global_id_set).is_empty()
+
+
+class WriteRaceChecker:
+    """Used for checking for overlap between access ranges of instructions."""
+
+    def __init__(self, kernel, callables_table):
+        self.kernel = kernel
+        self.callables_table = callables_table
+
+    @property
+    @memoize_method
+    def vars(self):
+        return (self.kernel.get_written_variables()
+                | self.kernel.get_read_variables())
+
+    @memoize_method
+    def _get_access_maps(self, insn_id, access_dir):
+        from loopy.symbolic import BatchedAccessMapMapper
+        from collections import defaultdict
+
+        insn = self.kernel.id_to_insn[insn_id]
+
+        exprs = list(insn.assignees)
+        if access_dir == "any":
+            exprs.append(insn.expression)
+            exprs.extend(insn.predicates)
+
+        access_maps = defaultdict(lambda: False)
+
+        arm = BatchedAccessMapMapper(self.kernel, self.vars, overestimate=True)
+
+        for expr in exprs:
+            arm(expr, insn.within_inames)
+
+        for name in arm.access_maps:
+            if arm.bad_subscripts[name]:
+                access_maps[name] = True
+                continue
+            access_maps[name] = arm.access_maps[name][insn.within_inames]
+
+        return access_maps
+
+    def _get_access_map_for_var(self, insn_id, access_dir, var_name):
+        assert access_dir in ["w", "any"]
+
+        insn = self.kernel.id_to_insn[insn_id]
+        # Access range checks only apply to assignment-style instructions. For
+        # non-assignments, we rely on read/write dependency information.
+        from loopy.kernel.instruction import MultiAssignmentBase
+        if not isinstance(insn, MultiAssignmentBase):
+            if access_dir == "any":
+                return var_name in insn.dependency_names()
+            else:
+                return var_name in insn.write_dependency_names()
+
+        return self._get_access_maps(insn_id, access_dir)[var_name]
+
+    def do_accesses_result_in_races(self, insn1, insn1_dir, insn2, insn2_dir,
+                                    var_name):
+        """Determine whether the access maps to *var_name* in the two given
+        instructions result in write races owing to concurrent iname tags. This
+        determination is made 'conservatively', i.e. if precise information is
+        unavailable (for ex. if one of the instructions accesses *var_name* via
+        indirection), it is concluded that the ranges overlap.
+
+        :arg insn1_dir: either ``"w"`` or ``"any"``, to indicate which
+            type of access is desired--writing or any
+        :arg insn2_dir: either ``"w"`` or ``"any"``
+        :returns: a :class:`bool`
+        """
+
+        insn1_amap = self._get_access_map_for_var(insn1, insn1_dir, var_name)
+        insn2_amap = self._get_access_map_for_var(insn2, insn2_dir, var_name)
+
+        if insn1_amap is False or insn2_amap is False:
+            return False
+        if insn1_amap is True or insn2_amap is True:
+            return True
+
+        return _check_for_access_races(insn1_amap, self.kernel.id_to_insn[insn1],
+                                       insn2_amap, self.kernel.id_to_insn[insn2],
+                                       self.kernel, self.callables_table)
 
 # }}}
